+++ udp-java/MTP/src/main/java/chat/dim/tlv/Length.java
@@ -32,7 +32,7 @@ package chat.dim.tlv;
 
 public class Length extends IntData {
 
-    public Length(byte[] data, int value) {
+    public Length(byte[] data, long value) {
         super(data, value);
     }
 
@@ -43,7 +43,7 @@ public class Length extends IntData {
         } else if (length > 2) {
             data = slice(data, 0, 2);
         }
-        int value = bytesToInt(data);
+        long value = bytesToInt(data);
         return new Length(data, value);
     }
 }
